add new src component and move the others, also made the dispatching actions for the api
