+++ index.js
@@ -1,23 +1,13 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-import { Provider } from 'react-redux';
+import FullWidget from './src';
+import {
+  addUserMessage,
+  addResponseMessage,
+  addLinkSnippet
+} from './src/store/actions/dispatching';
 
-import Widget from './src';
-import store from './src/store/store';
-
-const FullWidget = props =>
-  <Provider store={store}>
-    <Widget
-      responseMessages={props.responseMessages}
-      handleNewUserMessage={props.handleNewUserMessage}
-      profileAvatar={props.profileAvatar}
-    />
-  </Provider>;
-
-FullWidget.propTypes = {
-  responseMessages: PropTypes.arrayOf(PropTypes.object).isRequired,
-  handleNewUserMessage: PropTypes.func.isRequired,
-  profileAvatar: PropTypes.string
+export {
+  FullWidget,
+  addUserMessage,
+  addResponseMessage,
+  addLinkSnippet
 };
-
-export default FullWidget;
