+++ package.json
@@ -19,9 +19,9 @@
   ],
   "dependencies": {
     "immutable": "^3.8.1",
-    "lodash": "^4.17.4",
     "moment": "^2.18.1",
     "prop-types": "^15.5.10",
+    "react-immutable-proptypes": "^2.1.0",
     "react-redux": "^5.0.6",
     "redux": "^3.7.2",
     "redux-logger": "^3.0.6"
