+++ src/components/Widget/components/Conversation/components/Header/style.scss
@@ -1,4 +1,4 @@
-@import "../../../../sass/variables.scss";
+@import "../../../../../../sass/variables.scss";
 
 .header {
   text-align: center;
