+++ src/components/Widget/components/Conversation/components/Messages/components/Message/index.js
