+++ src/components/Widget/components/Conversation/components/Messages/components/Message/styles.scss
@@ -1,5 +1,5 @@
-@import "../../../../../../sass/variables.scss";
-@import "../../../../../../sass/common.scss";
+@import "../../../../../../../../sass/variables.scss";
+@import "../../../../../../../../sass/common.scss";
 
 .message {
   margin: 10px;
