+++ src/components/Widget/components/Conversation/components/Messages/components/Snippet/styles.scss
@@ -1,5 +1,5 @@
-@import "../../../../../../sass/variables.scss";
-@import "../../../../../../sass/common.scss";
+@import "../../../../../../../../sass/variables.scss";
+@import "../../../../../../../../sass/common.scss";
 
 .snippet {
   @include message-bubble($grey-2);
