+++ src/components/Widget/components/Conversation/components/Messages/styles.scss
@@ -1,4 +1,4 @@
-@import "../../../../sass/variables.scss";
+@import "../../../../../../sass/variables.scss";
 
 .messages-container {
   background-color: $white;
