+++ src/components/Widget/components/Conversation/components/Sender/index.js
