+++ src/components/Widget/components/Conversation/components/Sender/style.scss
@@ -1,4 +1,4 @@
-@import "../../../../sass/variables.scss";
+@import "../../../../../../sass/variables.scss";
 
 .sender {
   display: flex;
