+++ src/components/Widget/components/Conversation/style.scss
@@ -1,5 +1,5 @@
-@import "../../sass/variables.scss";
-@import "../../sass/animation.scss";
+@import "../../../../sass/variables.scss";
+@import "../../../../sass/animation.scss";
 
 .conversation-container {
   border-radius: 10px;
