+++ src/components/Widget/components/Launcher/style.scss
@@ -1,5 +1,5 @@
-@import "../../sass/variables.scss";
-@import "../../sass/animation.scss";
+@import "../../../../sass/variables.scss";
+@import "../../../../sass/animation.scss";
 
 .launcher {
   align-self: flex-end;
