+++ src/components/Widget/style.scss
@@ -1,4 +1,4 @@
-@import "./sass/animation.scss";
+@import "../../sass/animation.scss";
 
 .widget-container {
   bottom: 0;
