+++ src/store/actions/dispatching.js
@@ -1,0 +1,14 @@
+import store from '../store';
+import * as actions from './index';
+
+export function addUserMessage(text) {
+  store.dispatch(actions.addUserMessage(text));
+}
+
+export function addResponseMessage(text) {
+  store.dispatch(actions.addResponseMessage(text));
+}
+
+export function addLinkSnippet(link) {
+  store.dispatch(actions.addLinkSnippet(link));
+}
