+++ src/store/reducers/behaviorReducer.js
@@ -1,11 +1,11 @@
 import { Map } from 'immutable';
-import { TOGGLE_CHAT } from '../actions/actionTypes';
+import * as actionTypes from '../actions/actionTypes';
 
 const initialState = Map({ showChat: false });
 
 export default function reducer(state = initialState, action) {
   switch (action.type) {
-    case TOGGLE_CHAT: {
+    case actionTypes.TOGGLE_CHAT: {
       return state.update('showChat', showChat => !showChat);
     }
     default:
