+++ src/store/store.js
@@ -2,9 +2,9 @@ import { applyMiddleware, createStore, combineReducers } from 'redux';
 import { createLogger } from 'redux-logger';
 
 import behavior from './reducers/behaviorReducer';
-import messages from './reducers/messagesReducer';
+import messagesReducer from './reducers/messagesReducer';
 
-const reducer = combineReducers({ behavior, messages });
+const reducer = combineReducers({ behavior, messagesReducer });
 const middleware = applyMiddleware(createLogger());
 
 export default createStore(reducer, middleware);
